cfnlite->deps.py: Add dependency resolver
Similar to cloudFormation, cfnlite files allow users to reference other
resources using the `ref` keyword.

For this to work, references to resources need to be resolved.

This commit adds a dependency resolver which finds all the `ref`'s in the
cfnlite file and creates a topological ordering of the resources.

Resource building functions can then use this ordering instead of linearly
building them from the order entered into the cfnlite file.