cfnlite/securitygroups.py: Add support for SGs
This commit adds support for creating security groups.

cfnlite allows users to create security group egress and ingress rules for 4
predefined protocol: http, https, icmp and ssh.

If a user wants to create a custom rule, cfnlite will generate a standard http
rule and it will be up to the user to edit the generated CFN to point to the
correct port numbers.