cfnlite: Recursively update resources
Previously, we directly key into resource objects while generating the CFN
template. While this works with relatively flat resource types, CFN can be
pretty deeply nested (looking at you policy document) and that approach isnt
particularly scalable.

This commit implements nested update and find functionalities. The core reason
for this is that we can update values regardless of where they are in the
resource object.

As part of this commit we also update various parts of the existing code to use
this new functionality.