cfnlite: Ensure resources accept all properties
Previously, we directly keyed into the resource objects as we added new
properties. Due to how keying into object in python works (create or update)
we could build up the resource as we went along.

A recent change that recursively updates objects rather than directly keying
into them means that we no longer benefit from the python default create/update
behaviour anymore as the new functionality only updates, meaning that for it to
work the key must exist in the object.

We now change our "expectation" such that our defaults have all the possible keys
for every resource and then we delete any resource properties that the user did
not add via the cfnlite file.

Furthermore, we also define defaults for each resource that we want to keep
regardless of whether the user defines them or not, typically these are required
properties.