cfnlite->validators: Refactor prop validation
We're at the point of coalescing around a standard pattern for how we generate
CFN.

The basic property validation and ref resolution code is used again and again
for pretty much every resource so it makes sense to colocate this functionality
and share it across the library.

This commit creates this new validator file and removes any local definitions
of the same functionality in already existing resources.