cnflite->validators.py: Remove expects list check
Previously, as part of reference resolution, we passing a set containing the
name of all the properties that expected a list. This was so we could handle
resolving references which are inside a list.

However, we are passing in an object containing all the properly named and
assigned properties. This is because we do our ref resolutions after we've
validated, cleaned and correctly assigned any incoming property. As part of
this functionality we also find the value associated with the property name
we need to check references for.

These two facts combined means that we can just do an instance type check on the
value as if the property expects a list, the value will already by of type list.
And then we can branch off and check for references in that list.

This changes makes the code more robust and future proof. Currently, if the
property name was not in the expects list set and was in fact a list, our code
would not resolve any reference in that list.

This code would not have such a bug.