cfnlite/__main__.py: handle explain first
Previously, we were doing checks for all our cli args before checking if the
user passed the explain flag. The is generally inefficient and inflexible as
the explain flag allows us to exit early.

This commit defers any non-explain flag checks until we confirm that the
explain flag has not been passed