cfnlite/__main__.py: fix explain flag bug
Previously, when `--explain resources` was passed into the CLI, instead of
returning a list of the resources that cfnlite supports, nothing happened. This
is because the output was not being printed.

This commit fixes the bug but also adds some aesthetic changes to how the output
is presented to the user.