cnflite/__main__.py: turn input file to a flag
Previously, the input file was a required parameter which made using the
`--explain` flag a pain as users would still need to pass a file even when they
only care about the resource property information.

There are two main solutions to this, either create a subparser or make the
required arg into an optional flag. We don't really have enough commands/options
to justify a subparser so this commit changes the input file into an optional
flag.