cfnlite->deps.py: clean value before checking
Previously, as part of building the dependency tree, we didn't remove whitepaces
around incoming values when searching for a `ref` flags. While this doesn't
break everything, it makes the system significantly more brittle.

Furthermore, we did not ensure that the value case was lowered before looking
for the `ref` keyword which would break cases where a user puts `ref` with an
uppercase `R` instead of the lowercase which the current implementation expects.

This commit fixes both these bugs.