chore: Make referencing an explicit keyword
Previously, we used a the word `ref` to signify resource referencing in cfnlite.
As we begin to add more keywords we would like it to explicitly clarify what is
a keyword and what is not - especially as most keywords can show up in other
words e.g. referral which could lead to weird bugs.

Initially, the idea was to do the similar thing to CFN which uses `!` in front
of keywords e.g. !Ref. However, it turns out that putting `!` in front of a
token is invalid YAML[1], do AWS must have a custom YAML parser for this to work.

We, currently, have no intention of writing our own parser so in order to get
around the issue we have decided to put the `!` after keywords, as this should
make keyword matching/parsing more accurate regardless of the context.

We did consider using other special characters (# $ ? @ % & *) but it turns out
putting any of those in front of a token is similarly invalid YAML.

[1] https://stackoverflow.com/a/9703472