[bug] cflite/__main__.py: fix name check
Previously, we directly keyed into the parsed cfnlite file to check if the file
has the required `name` property and if not, raise an error. The problem being
that if the name does not in fact exist, it caused a KeyError instead of the
expected ValueError, so the program blows up in a way that is unexpected.

This commit fixes the issue by doing a pessimistic check into the parsed file
using `.get()` instead of directly indexing. This allows us to error handle
the way we initially intended.